backend/requirements.txt
fastapi
uvicorn[standard]
openai
requests
python-multipart
python-dotenv
boto3
celery
redis
sqlalchemy
psycopg2-binary
alembic
websockets
python-multipart
wave
numpy